BL-14402 refactor langtagProcessing.ts
